Move `with_cleanup` to `coordinate::cleanup`
